addition speed detection with ld (HL)
use two more state machines within the the speed detection
algorithm. There is now a state machine which detects the
ld hl opcode and one which detects the SET 0, (HL) opcode.

This can follow the ROM if it is setting bit 0 in the key1
register of the gameboy via the HL CPU register.